Add wfv to calls_correct and prove cheat in App case
The theorem statement is now annoyingly intricate, since certain
assumptions are required for parts of the conclusion, but are too strong
for the inductive hypothesis in general. The alternative would be to
prove two theorems, but they would have duplicated a lot of proof.

The ev_app cases have not been updated (broken and commented out).