Fix a bug in the exhaustiveness checker
Not a soundness bug obviously (since it was verified), but one where it
would fail to detect clearly exhaustive matches. Ideally we will one day
have a proved-complete checker.

Yong Kiam and Scott suggest this may have been fixed in the past and
therefore possibly not carried along in some merge or refactor.