Add word_simp which will contain simple wordLang phases
The composition of bvp-to-word is complicated and difficult
to navigate, but we want to add a few simple optimisation
passes that run right after translation into wordLang and
before register allocation.

The purpose of word_simp is to be an understandable entry
point into this part of the compiler. The correctness
requirements for word_simp are strict and simple:

    evaluate (prog,s) = (res,s2) /\ res <> SOME Error ==>
    evaluate (word_simp$compile_exp prog,s) = (res,s2)

    extract_labels (word_simp$compile_exp p) = extract_labels p

The current implementation of word_simp is the identity
function, but will be expanded to include at least:

 - clean up of wordLang programs: delete Skips and make Seq
   associate to left so that pattern matching is easier in
   other phases. I'm working on this.

 - combine certain consequtive If statements (issue #103)
   I'm working on this.

 - move load instructions upwards (issue #118 ?) Someone else
   might be working on this.