Re-organising a bit
Removing the flat_tenvC_to_ctMap constant makes several theorems about
type definitions building good environments obsolete, and I'll have to
re-do them. But it is much better to existentially quantify the ctMap,
rather than try to build it out of the range of the lexical constructor
map (which is why that had to be an alist instead of fmap before).