Make word_simp optimise If statements
It now:
 (1) left-associates all Seqs and removes all Skips
 (2) fuses some consequtive If statements

Needs doing:
 (3) constant propagation
 (4) deadcode elimintion to remove residue of (2) and (3)

Closes #103

Problem: I cannot prove that the syntactic requirement is met.
The syntactic requirement states that the labels are unchanged.
They are morally not changed, but their order in the accumulated
list may change. I think the syntactic constraint should be
relaxed. Could it be a bag instead of a list?