Finish bvl_inline implementation and proofs
The implementation works by walking the code table from top
to bottom. The loop inlines to-be-inlined functions and then
checks whether the resulting code is small enough (and not
recursive) to decide whether it should put the function's
body into the map of to-be-inlined functions.

The proof is interesting because it uses a semantic relation
rather than the usual syntactic approach.

Progress on #51

Thw issue will be closed when the magic number in bvl_to_bvi
has been made into compiler configuration. Currently a
function is small if its size is less than 3. (This
optimisation is turned off by setting the limit to 0.)