Work around parser inefficiency in cmlPEG
Issue is to do with repeatedly type-checking large terms (in this case,
that have been antiquoted into the parse). I think this can be fixed,
and I will work on this in HOL. In the meantime, the workaround is to
build the term explicitly and to not bother with the parser at all.