Improve specifications of the basis program
It is always better to put pure facts as assumptions than in the
precondition. When doing a proof, xpull will always be called first,
extracting pure facts from the precondition. With these facts not in the
precondition, the specification therefore have chances to be directly
used without using the frame rule.