Start a refactor of environments
This is a start on #126, but only for the operational semantics. Next up
is the type system.

Previously, declarations and top-levels carefully kept the constructor
environment and value environments very separate: the first was always
returned along with the state, whereas the second only when returning an
Rval result. This was required due to the bad way that the old type
soundness proof (pre-fun-big-step refactor) calculated the constructor
identities map from the environment. Now it existentially quantifies
one, so we get back flexibility in the environment design.

Now, both decs and tops take in an environment record with two fields,
one for values and one for constructors as before, but they return the
same type in their results. This makes the semantics much cleaner.

I also changed identifiers and environments to support arbitrarily deep
nesting, looking forward to a later time when tops and decs can be
merged, and nested modules supported. This also cleans things up by
introducing some more uniformity.