Add some automation for cf_examples_pmatch.sml
This automation proves the theorem equating a PMATCH_ROW based on a list
insts to another PMATCH_ROW based on a tuple corresponding to the list
of pat_bindings. I am not sure how best to use it, though. I included an
example usage, but it does not yet solve the goal.