Fix type soundness invariants
For #126.

Decided to move the file out of Lem, because it needs mono theorems from
environmentPropsTheory to define the type_v inductive relation. Also Lem
was generating some weird code for (fun x (y,z) -> ...) that was messing
up the monitonicity. Since this isn't in the source semantics, it
doesn't really need to be in Lem, and there's no going back on the
decision to not use Lem for IL definitions given the number of them that
we have.