Add to_dataBootstrapTheory
This theory will evaluate the to_data compilation for the various
compiler progs (for each target). (Without incremental compilation,
these all need to be done separately in full.) Currently, it only does
the prog declaring the inferencer, since the rest of the compiler is
still being translated.

Evaluation is done in stages with saved intermediate theorems, for
better interactive use and finer grained timing. However, it is fairly
naive, and doesn't currently include many tricks to make evaluation
faster.