Make to_dataBootstrap compile compilerProg
Plus a few tweaks and fixes.

This takes about 10 minutes, most of which (7 min) is in bvl_to_bvi,
probably due to its use of slow appends (which can be fixed).

However, the parser and inferencer are now not included, because the
translation is not set up to include both the parser/inferencer and the
compiler. I presume this will be fixed in the translation at some point.