Make clos_to_bvl sort the code-table entries
The sorting ensures that code for a closure appears
immediately before any clos_call entry for the body
of the closure. The hope is that this will mean that
the tail-call from the code table entry in the closure
can in most cases be optimised to a Skip in LabLang
and then removed (see #131).