Progress on x64 bootstrap
Stopped before remove_labels_loop. For some reason x64_enc is not being
evaluated properly (but my state could be bad).

There are a few functions in the backend that are essentially MAPs but
are not defined that way; if they were characterised as MAPs, they would
be more amenable to parallelised evaluation.