Improve x64Bootstrap slightly, and make some progress
Reworked the parallelisation so that it keeps all threads busy as long
as there is still work to be done (modulo a configurable chunk size).
Cleaned the code up a little.

Parallelised filter_skip (by showing it is a MAP).

Looked at the last bit of lab_to_target (remove_labels_loop) and
started devising a plan for evaluating it.

Showed that x64_enc is not being evaluated on all terms, even if using
x64_compileLib.eval. Not sure if this is a problem.