Make progress on x64Bootstrap
Up to the second run of enc_secs_again. Each run takes about 1s per
prog, and there are currently 1800 progs, but that is expected to grow
(maybe double) when the rest of the compiler is translated. I'm not sure
how much thought has gone into optimising the assembler (lab_to_target);
a faster and/or more parallelisable algorithm would probably help, but
I'm not sure if it exists.

Also, put compilerProg and x64_config in base heap.