Write last part of x64BootstrapScript.sml
Introduce a cheat for the all_enc_ok check, rather than evaluating it,
since it is slow and unnecessary. Then, produce the final theorem (with
concrete bytes on the right-hand side). It takes a long time to evaluate
FLAT on these bytes (which start as a list of lists) - there may be a
better conversion for doing so than the one I am using. Also, currently
listLib.FOLDR_CONV is broken (although this commit uses it, so will fail).

The bootstrap_thm is the only exported theorem. Writing the bytes to a
.S file can happen in a later theory, or could be added to the end of
this.