Make FromList take length as argument in BVI
FromList's implementation in data-to-word will be simpler
if the length of the list is already provided. This commit
adds a stub to bvl-to-bvi. The stubs computes the length
of the list.

There is one cheat bvl_to_bviProof that I just can't seem
to see where a part of that goal comes from.