Reduce lab_to_target exit checks to one (big) cheat
The cheat is that when code_alignment is not 0, then
  - every label is length 0
  - every line's annotated length is 0 mod 2 ** code_alignment
  - every line's actual length (i.e. length of bytes) similarly
for the code before padding.

Most of this would require knowing asm_ok (in combination with enc_ok),
so asm_ok would need to be pushed up the compiler. However, for jump
offsets, it's not obvious that this is going to be true: there might be
an intermediate jump offset (before reaching the fixpoint) which is not
asm_ok. I'm not sure yet; need to look into it more.