Attempt to remove lab_lookup from exit checks
Good news is that this is possible (only one cheat remains).

The bad news is that this uncovered a mistake in all
semantics from BVI down to LabLang. In each language, the
primitive for creating a CodePtr or Loc ought to check that
this location exists in the code table, and fail if it
doesn't. Currently, BVL is the only language to get this
right. Making the semantics do the right checks will affect
approximately half of all compiler proofs. Fortunately,
the repairs ought to be trivial in all instances, since
the verification of Call in each transformation requires
the code tables to be sufficiently related.