Fix word_allocProof and make it more debuggable
Debuggable: I added a lot of semicolons to word_allocProofScript.

I think it's good practice to put a semicolon after each top-level
declaration so that a failure in one declaration still stores all the
values bound by the declarations before it. (This is part of the
Convetions page on the wiki.)