Make bvl_handle smarter w.r.t. Seq when env is []
This optimisation only applies to top-level Seq commands
in code-table entries with arity zero. This optimisation
aims at making bvl_handle more efficient for the main
expression that contains all of the declarations. It
ought to be more efficient to EVAL and also produce
code-table entries that are near trivial for the
register allocator in most cases.

See comment on #133.