Define full_normalise (and state its correctness)
The correctness statement is in a comment, but can be used as a
starting point for an experiment trying out whether it can be used
directly with the cf function. In order to use it, sound_def ought to
change to make use of norm_rel, norm_res_rel, norm_state_rel etc.