Update the examples and the tutorial to use [full_normalise]
In theory, the source program could be rewritten without the hand normalisation,
and the proof be identical. Here we needed to update some proofs as the
automatically normalised programmed differed a bit due to the order of
evaluation. The hand normalised programs corresponded to a left-to-right
evaluation order, while it is actually right-to-left.