Prove more about top-level FFI semantics
Proved the following property: If the io_events is not changed and the
final_event is still NONE, then the entire ffi state is not changed.

Also, encapsulated this and similar properties in a new io_events_mono
predicate. If more need to be added later, this should make maintenance
easier.