Fix semantics heap deps to include backendTheory
Otherwise you can change wordLang and not have the semantics heap
rebuilt...

This will cause any change in backend to cause rebuild all of semantics.
This is what should happen given the presence of a heap in semantics,
but I'm not sure it's ideal.  Perhaps the heap in semantics can be
dispensed with entirely...

The issue is that there are many independent dependency "strands" from
backend into semantics, but having a heap forces all the strands to go
through one bottleneck. Also, it's not clear that any individual theory
in semantics really inherits that many theories in backend, perhaps
suggesting that the overhead of having to load them when the semantics
script runs won't be so great.