Add operations; complete coding for cat file sys
Operations are open-a-file (getting a fd), read-a-char-from-a-fd, and
close-a-fd. There are various error conditions (some of which might map
onto exceptions). For example, opening a non-existent file might
reasonably throw something. On the other hand, attempting fgetc when at
eof plausibly should return NONE instead of SOME char at the client
level (as well as in the underlying model).

These operations now need to be lifted into values of type ffi_next,
which will operate over encoded file-system values.