Make mem loads use one fewer instructions
The default is to do two shifts to extract the address from pointer
with information in the lower bits. This commit makes the compiler
produce code with only one shift if there are enough pad_bits. There
needs to be two pad bits for 64-bit archs and one pad bit for 32-bit
archs.