Add sugar for boilerplate function definitions
The otherwise_homomorphic function takes a function
definition over a (mutually) inductive datatype
that is possibly missing some clauses, and generates
a definition covering the missing clauses, under the
assumption that the function simply distributes
over the constructor in the missing clauses.
Still rather crude; for an example, it assumes
that patterns are not nested.