Start Updating modLang and conLang for nested modules
The env refactoring assumes a uniformity of module paths that will
naturally support nested modules, but modLang and conLang have stuff in
them that isn't flexible enough, and makes it hard to port them to the
new namespace structure.

In particular, modLang groups top-level definitions into prompts. A
prompt used to have an optional module name, but that's not really a
good idea, since modules are supposed to be compiled away for modLang.