Fix catTheory for changes to string primitives
Since catTheory has its own copy of the string module translation, we
need to change the primitives here too, and translate explode in terms
of them. Then one of the spec proofs needed a bit of maintenance.

explode has a side condition, which is provable. But calling
update_precondition on it breaks the CF framework. This needs further
investigation (but can be postponed).