Make [cfAppLib.app_of_Arrow_rule] more generic wrt. the ffi type
[app_of_Arrow_rule] takes a theorem of the form [(A --> B) fv f], and turns it
into a theorem of the form [app p fv [xv] ...]. The first theorem does not
involve any ffi type, while the second does: [p] has type [α ffi_proj] for some
type [α], that the rule has to insert.

Previously, the rule always inserted the ['ffi] type variable as [α]. It now
requires the ffi type to be put there as an argument, and xapp/xspec provides
a ffi type fetched from the goal.