Fix translator (hol2deep) for partially instantiated constants
In doing tweaks for the abstract mode, I had switched this to
same_const, but when the same constant needs different translations at
different types, this is not precise enough. Now (as before abstract
mode) we only match where a previously translated constant can be
instantiated into the desired constant.