Continue updating data_to_wordProps
Including with help from Magnus.
The tagging scheme seems workable.

Changed the representation of byte arrays slightly (the last word needs
to be zeroed out beyond the last byte) so we can reuse the word_cmp_loop
test when comparing byte arrays (which compares a whole word a at a
time).

There is cheat in word_eq_thm about using word_cmp_loop. And the
compiler needs to be changed for RefByte.

(issue #100)