Abstract the getArgs spec out of commandLine_fun
The idea is to reuse ffi_getArgs for both the CF view of the FFI and the (raw)
CakeML view. We are attempting to establish a clean workflow for verifying
whole programs with CF. Defined encode/decode functions into and out of the
(CF) ffi type. Also, cleaned up various things.