Start updating data_to_wordProof
Fix the header produced by RefByte's implementation. It was easier to
pass a number containing the upper bit (or not) for the byte header, and
Or the lower bits (7w) separately within RefByte, because then the
header tag argument to RefByte_code has a representation as a dataLang
Number (can't be confused for a pointer).

The proof is now broken at the point where the implementation needs
work: either Replicate needs to And the final word with a mask, or
RefByte needs to call Replicate as a non-tail call and then do the mask
afterwards.