Reuse vector type definition from Konrad's regexp stuff
This is to avoid the awkward duplication of the vector type with the
current setup. Not an ideal solution; in particular the dependency of
mlvector on regexp_compiler is bizarre. Long-term I guess the vector
definition ought to live in a theory of its own, on which the basis
library and regexp_compiler will depend.