[ffi] Flatten command line arguments.
Command line arguments to any program following the C calling convention
comprises of 2 parts:
* the number of arguments(including the program name).
* the arguments themselves(as a vector - an array of character arrays)

This patch stores these values in a global address and then flattens the
vector to a byte array, where each argument is separated by a '\0' and
the byte array ends with a series of 3 '\0' characters.

For example, the vector:
argv[][] = {
  "progname",
  "arg1",
  "arg2",
  "arg3"
};

Will be flattened to look like:

char * args = "progname\0arg1\0arg2\0arg3\0\0\0";

The flattening of the vector is done in C.

The assembly side of things is only done for x86_64 for now, Need to implement
arm6, arm8, riscv and mips support.