Create presLang, the presentation language, and ast -> presLang function
Makes presLang, which is intended to capture all features of several
intermediate languages. In particular, it will capture all those that
are mainly functional in style. As such, it might be extended with
several constructors.

presLang has only one data type, exp, as it lacks semantics. The purpose
of preslang is for presenting tha program structure as a tree, not to
evaluate it.

This change includes some test cases for presLang, in testScript.