Simplify the quicksort invariants
Instead of explaining how to break a list apart by lower and upper
pointers, only ever have the pieces and append them together, and
define the shallowly embedded lower and upper in terms of the length of
the pieces.

partition's spec and the proof still need updates.

Also, we zip the shallowly embedded and deeply embedded values when
checking for a permutation to avoid requantifying the shallow side to
check for sortedness in the post-condition. This should avoid any
funniness if the relation a is not suitable deterministic.