Add xopb and xopn and remove cheat in mlcommandlineProg
The cheat was for a stronger spec for tabulate which propagates the use of the
function argument correctly. To be able to manually prove a CF spec about this
translated tabulate function, we needed to manually normalise its code and also
use PRE instead of _ - 1, since the translator's apprach to subtraction
produces non-normalised code.