Start on wrapper interface for expressing FFI functions
In this approach, an ffi part has the following type:

  ('a -> ffi  # (* encode *)
   ffi -> 'a  # (* decode *)
   ( string # ('a -> word8 list -> 'a option )) list
     (* named ffi operations on 'a *)

The hope is that this offers a more structured, modular approach, from
which the CF IO and proj things can be defined systematically.

This commit defines the helpers IOx, mk_proj1, and mk_proj2, and defines
ffi parts of the above type for stdin, stdout, commandLine, and rofs
(file system). The latter is mostly a copy of catfileSystemTheory (which
will be deleted eventually).