Avoid feof in ffiisEof function in basis_ffi.c
feof tests if the end-of-stream marker is set, which is not the same as
whether the stream offset is actually at the end. So, resorting to
fgetc/ungetc to test eof. (For STDIN, we use a "read_failed" flag
which gets returned with the getc call instead of a separate isEof
function.)