Use set_grammar_ancestry to purge grammars of llist and sptree
Even if one is committed to including every possible ancestral theory
from HOL (via misc and/or preamble), one need not let their parsing
context influence the current one.

Currently, it's clearly far from ideal to depend on regexp_compiler
for the definition of the vector type; doing so brings in
balanced_map (which has its own notion of fromList).