Verify get_file_contents
Prove that if you repeatedly call inputLine and accumulate the resulting
strings in a list of strings, you get the right thing. The
post-condition is in terms of the linesFD function from the rofs
definition. The definition looks ok, but I think it would be best to
eventually prove that linesFD implements a more declarative
specification of what the lines of a file are.