Make semantics' notion of parse deterministic
The old text had something like

  some pt.   ptree_fringe pt = MAP TOK toks /\ ...

There will be multiple parse-trees satisfying that constraint because
this doesn't constrain the location information that appears in the
parse-tree.

The new text uses real_fringe and makes the parse function take what
the lexer actually returns, which is a list of tokens coupled with
their locations.