Get sort down to a single cheat
The cheat is for a presumably false goal of the form

ARRAY _ array_v ==>> emp

I think this is happening because the sort program allocates an array
internally, but the top-level spec doesn't mention the array in the
post-condition, since it isn't relevant. (In fact, mentioning it would
arguably violate abstraction by revealing how sort does its work.)

In a usual separation logic for a C-like language, you either would free
the array in the sort code, and then it wouldn't appear in the
post-condition, or if you don't free the array, it must appear in the
post-condition so that ownership transfers to the calling code. However,
we have a GC, and so I want to act as though I've freed the array since
it is not returned from the function. I'm not sure how to solve this
problem. It's possible I've just mis-instantiated one of the existential
quantifiers of hprops.