Add an FFI operator token and concrete syntax
The FFI name gets enclosed in #(...), so that #(read) will map to FFIT
"read", for example. The lexer_fun specification allows for
whitespace (but not newlines) within the parentheses. This flexibility
is probably completely unnecessary, but the definition illustrates how
we might play with what's possible here.

I'm confident that #( cannot appear in real SML programs (Poly/ML
does accept stuff like #(* foo *) 1 (2,3); evaluating to 2).