Put FFI tokens into the grammar and adjust cmlPtreeConversion
Adjustment to the latter turns what is parsed as

  ((#(read) · arg1) · arg2) · arg3

(with · as the infix function application operation) into

  Op (FFI "read") [arg1; arg2; arg3]

Of course, the current design of the semantics insists that there only
ever be one argument to FFI ops, but the syntax need not know about
that.

Along the way delete a sanity check lemma in the latter theory that
was tedious to write and run. The test-cases for the PEG/conversion
code should test this stuff adequately.  The totality proof for
cmlPtreeConversion in semantics/proofs goes through unchanged.