Use a .holpath file to make CakeML relocatable
If using a recent HOL (post 66ac9175f) CakeML object files will not
need to be rebuilt if CakeML or HOL move around.  Heaps and .HOLMK
directories do need to be deleted and recreated though.  To avoid
the rebuilding of heaps prompting the rebuilding of other files, use
Holmake --relocbuild