Make List.tabulate tail rec. and apply in asc. order
The previous implementation was both non-tail-recursive and applied the
function in reverse order (but none of the specs for it check that:
they assume the function is relatively pure).

Note: to use the simple translator spec for List.tabulate, one must use
xapp_spec tabulate_v_thm because bare xapp will probably pick the more
complex CF spec by default.