Rework mipsProg translation with Yong Kiam
Following the inlining strategy that Yong Kiam is experimenting with on
the inlineenc branch for x64_enc. (I commit this to master now because
mipsProg was broken anyway.)

The strategy:
Rather than translate the encoder and all its auxiliary functions,
which are more general than required for encoding CakeML's ASM, we
inline parts of the encoder and ASM translation to get a specialised
version of the encoder for CakeML.

This makes the translation shorter and simpler (there are no
side-conditions any more, for example).

It's a somewhat manual process, and the resulting code could probably be
improved even more.