Prove new side condition in source_to_mod translation
The astOp_to_modOp function is partial, but used in a good context.
I think this was missed in ee8866480e704266e3b5b70d735bc50eb36f012d.

I think this should remove the need for the cheat in compiler_x64Prog,
but have not tested this yet. 476584d728724556f1f128e74aa97c01baebc3db
claims that cheat is due to EqualityType, which I think is bogus.