Do some timing to compare evaluation strategies on to_data
Specifically, compiling helloProg to dataLang using either the original
bootstrap method, or pure computeLib on the whole task. It seems the
bootstrap method (which includes parallelism) actually does do better.
So the overall slowdown compared to computeLib on the whole compilation
is probably coming from another phase of the compiler.