Clean up wfFS and the CARD bound
It turns out wfFS is already part of ROFS, so did not need to be
included explicitly in specs. Also, the CARD assumption would be better
phrased (and is now overloaded as) hasFreeFD - which shouldn't be part
of ROFS since one could have used all the FDs but still have a fine
filesystem.