Extend bvi_tailrec analysis with type contexts
Implementation changed to allow for a slightly larger class of
expressions. Rewrites

> fun f x = ... ; f x' + y

to

> fun f x acc = ... ; f x' (y + acc)

Previously y was required to be integer constants or operations on
these. These changes should at least allow y to contain variables
so long as they are arguments to the function f.

Note: adjusting the theorems, so this commit upsets the proofs a bit.