Prepare to fuse clos_remove and clos_annotate
All of the information that clos_remove propagates in its pass is
available in clos_annotate. This commit starts an attempt to
incorporate the optimisation performed by clos_remove into the first
annotation pass in clos_annotate.

This commit doesn't yet change the compiler but rephrases part of the
clos_annotate proof to fit better with clos_remove.