Implement and verify Let and Letrec opt in clos_annotate
The optimisation is essentially the same that exists in
clos_remove. This suggests that clos_remove should be
removed from the compiler.

This commit leaves a cheat about evaluation of `pure`. It
also leaves a few cheaps about syntax at the bottom of
the file. Those theorems need updating based on similar
theorems in clos_removeProof. This commit does not remove
the files associated with clos_remove.