Reorganise BVL optimisations in prep for Install
What used to be `optimise` in `bvl_to_bvi` is now part of the last
part of bvl_inline. This has no impact on the what optimisations are
applied or their order. It is just a different way of organising the
compiler implementation around BVL.

The bvl_inline compiler now returns its state.

This commit does not update the proofs.