Update compilationLib after bvl_to_bvi reorg
Just calling eval directly on the whole thing now, because it looks like
the MAP in optimise, which was previously parallelised, has disappeared.
This probably makes evaluation slower, but hopefully not too much.

The inline step was also previously timed separately. Now it is folded
in with the rest of to_bvi.