Add support for resuming monadic translations
Adds functionality to export and import translations performed with the
monadic translator. State is automatically exported at calls to
`export_theory ()` just as with the standard translator, and can be
resumed by calling `m_translation_extends "..."`, which restores the
monadcic state and recursively calls on `translation_extends` from the
standard translator.