Add Rat module (rational values) to basis
The new module is placed after mlintProg and before mlcharProg.  The
script teaches the translator how to translate operations such as
fromInt, +, -, *, /, <, <= for reals. It proves that REAL_TYPE is an
EqualityType, which is possible because reals are rational numbers
here.

It does not teach the translator how to translate HOL's rat type. The
theory can be updated at a later date to support HOL's rat type.