Define forwardFD, linesFD, STD_OstreamFD
And maybe some other constants. Prove various properties about them.
Update proofs that were using bumpFD (and bumpLineFD, bumpAllFD) to use
the forward variants instead, since these formally ignore numchars -- I
realised that numchars should _only_ be used for the read system call.
At the STDIO level it is better to leave it untouched.

STD_OstreamFD is my workaround for the fact that the datatypes seem to
allow reading from stdout. The condition ¬STD_OstreamFD fs fd states
that fd is not a file descriptor for stdout or stderr, so reading from
it won't mess with outputting.