Update grepProg for fsio
The presentation of the grep semantics is changed rather significantly:
rather than producing a pair of strings (the output to produce on
stdout/stderr), I phrased the semantics as an fsio-updating action. This
makes parts of it easier to read, at the expense of a perhaps surprising
FOLDL. Also it means more lemmas are required about grep_sem (that it
doesn't modify the fsio state apart from the output streams). I think
this fit the proof better, but there may still be nicer ways to do
things.

In updating the proof, also switched to using xlet_auto wherever
possible, which made things much nicer.