Make a proof of concept of choreographic compilation for CakeML
We start from the most stupid choreography language of all time:
a choreography here is simply a list of handshakes, as in

 A --> B
 B --> A
 C --> A

project this to a parallel composition of endpoints like:

    (A, Send to B; Receive from B; Receive from C)
 || (B, Receive from A; Send to A)
 || (C, Send to A)

in an intermediate language that's like an old-school process algebra
in the CCS tradition. Semantics is synchronous for convenience;
to make it interesting we want asynchrony eventually.
We prove that the endpoint projection of a choreography is always
deadlock free.

Then compile the endpoints (here A) to CakeML programs like

 let val _ = ffi_send(B) in
  let val _ ffi_receive(B) in
   ffi_receive(C)

and prove that when run with an FFI oracle whose state is

  (B, Receive from A; Send to A) || (C, Send to A)

and where FFI next is just the transition relation from the intermediate
process algebra, the CakeML program always evaluates to a value.

In other words, we lift the deadlock-freedom-by-design of the
choreography language, down to failure-and-divergence-freedom of the
CakeML program that implements each endpoint (provided it runs in parallel
with programs implementing the other endpoints).