Rehaul bvi_tailrec and add ListAppend support
Previously, bvi_tailrec would perform two very similar phases:
1. Attempting to rewrite function tails with operations occuring in tail
   position.
2. Rewriting the entire function body with the first operation from (1) that
   succeeded.

The rewrite itself performed various things such as changing
associativity of operations and utilizing commutativity when possible.
The process was hard to follow and did not turn out well when ListAppend
(a non-commutative operation) was added to the mix. The procedure has
now been split into a few stages. In addition to being easier to follow,
the optimization can now be made to fail earlier if one of the stages
cannot be performed:

1. Check if function has the right recursion pattern at all, else fail.
2. Attempt to rotate all operations in tail position so that
   applications are right associative.
3. Attempt to move the recursive call to the leftmost position in the
   tree inorder using commutativity (optional).
4. Check if there are any suitable tail positions that can be rewritten
   if the above succeeds.
5. Rewrite the function using the operations from one of these (note
   that at this stage most rewriting has been performed by the earlier
   stages).