Add support for reading string terms from processes
Adds support to preamble for executing external processes and
capturing their stdout output as string or mlstring HOL terms.
The functionality is implemented using fork/exec and resembles
the PolyML implementation of Unix.execute from SML basis.
Commands are searched for in the directories specified in the
$PATH variable. Additionally, processes can be started with a
specified working directory.