Prove semantics thm for clos_to_bvl
The new semantics proof has an intentionally modular shape: I state
the shape of the simulation relation between evaluate functions and
prove that semantics is equal for any such simulation result. I then
separately show that the actual simulation between evaluate functions
fits the required shape. The hope is that this is easier to update
when the evalaute relation changes with time.

This commit does not include the (semantics-level) composition of the
separate internal closLang phases (clos_mti, clos_number,
clos_annotate etc.), i.e. this is just for the function that turns
closLang programs into BVL programs.