Make extract_fs restore original numchars
Use this to rework call_main_thm_basis (now
whole_prog_spec_semantics_prog) to a theorem that packages up the
user-proved assumption behind a constant, whole_prog_spec, whose main
argument is a post-condition on the final file system state.

By sticking to the original numchars, we hope to easily be able to use
things like (=) add_stdout for this post-condition.