Set up compiler for Tiny (and change link_reg for Tiny)
By changing the link_reg to 0 I could make the name permutation
function the identity, i.e. the names on Tiny correspond directly
to the naming convention used internally in the CakeML compiler.

There is a problem with the set up: I had to cheat a lemma about
the ranges of offsets and valid immediates because they weren't
large enough for the current encoding of asm into Tiny. I think
the asm-to-Tiny encoding should be modified.