Add some very simple benchmarks
Allocate an array of size 1000 and then iteraterate over it assigning to
it and the indexing it. Also add a dummy version that does all of the
looking and arithmetic but doesn't access the array. This is because
array accesses in this setting are very cheap and we need to subtract
out the time for array accesses.

On my machine, I get the following benchmark numbers, which means that
our arrays have poor performance.

CakeML : .9s (2.8s array - 1.9s dummy)
poly5.7.1 default config : .19s (1.5s - 1.3s)
mlton 20130715 : .17s (.48s - .31s)
ocamlopt 4.06.0 : .43s (.73s - .30s)
ocamlc 4.06.0 : 7s (17.1s - 10.8s)

I intend to add a few more benchmarks with sptrees and balanced binary
trees eventually.