Add an unverified sexpression AST printer
This change adds an unverified sexpression printer for
the CakeML AST, which enables bootstrapped compiler builds.
The translated compiler AST is printed as sexpressions
and then fed into a prior built compiler. This approach
reduces build times and has a lower memory footprint,
as the compilation process is not carried out in logic.
However, it does not provide the same correctness
guarantees and will fail if the compiler is implemented
with features newer than the built compiler.

The change also adds two compiler flags: `--exclude_prelude`
and `--skip_type_inference`. These toggle certain
behaviour off in the compilation process. They are
necessary when building the compiler using this approach
as the translated AST already includes the prelude
functions and will fail type inference checks.