Change how the axioms are defined
Inserting the axioms as terms baked using abbreviations leads to
excessive register allocation times. The expressions, INFINITY_AX in
particular, become huge.

This addresses the problem and implements a few monadic functions which
creates terms involving quantifiers, booleans, et cetera (using nothing
but functions and the constants = and @).

Note that INFINITY_AX is still broken. I suspect that the defnition of
surjectivity, injectivity -- or both -- is being applied in a bad way.