Make translator's approach to ind proofs more user friendly
The translator now prints (what I hope is) helpful messages when an
induction proof fails. The user is advised to prove the required
induction theorem and toggle a switch that makes the translator not
attempt the induction proof in the future (removes error message).

If a user fails to prove a dangling induction goal, then the
translator stops with an error either:
 - when translating something that wants to use the bad v thm, or
 - when the translator's state is packaged up by export_theory.